Tweak link-following behaviour.
The philosophy is that RET should do what org would do on a link; this
concern overrides the fact that dired and Buffer-menu-mode use switch
to buffer in current window on RET. Instead we use, currently, '.' for
current window. Like dired and Buffer-menu-mode we use 'o' for other
window (which may well have the same effect as RET).